# Conflicts:
#	src/Karteikartensystem/1. Code/Ansicht/WindowsFormsApp1/App.config
#	src/Karteikartensystem/1. Code/Ansicht/WindowsFormsApp1/Fenster/StapelUebersichtView.Designer.cs
#	src/Karteikartensystem/1. Code/Ansicht/WindowsFormsApp1/Fenster/StapelUebersichtView.cs

+++ src/Karteikartensystem/1. Code/Repository/Repositories/IKarteRepository.cs
@@ -9,9 +9,9 @@ namespace Repositories
     public interface IKarteRepository : IRepository
     {
         bool KarteHinzufügen(Karte karte);
-        Karte[] GetAlleKartenEinesStapels(int stapelId);
-        bool KarteLöschen(int id);
+        Karte[] GetAlleKartenEinesStapels(long stapelId);
+        bool KarteLöschen(long id);
         bool KarteAktualisieren(Karte karte);
-        int AlleKartenEinesStapelsLöschen(int stapelId);
+        int AlleKartenEinesStapelsLöschen(long stapelId);
     }
 }
